src/main/java/net/naonedbus/schedule/CommonScheduler.java
/**
 * <p>
 * Copyright (C) 2011 Romain Guefveneu
 * </p>
 * <p>
 * This file is part of naonedbus.
 * </p>
 * <p>
 * Naonedbus is free software: you can redistribute it and/or modify it under the terms of the GNU
 * General Public License as published by the Free Software Foundation, either version 3 of the
 * License, or (at your option) any later version.
 * </p>
 * <p>
 * Naonedbus is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
 * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
 * General Public License for more details.
 * </p>
 * <p>
 * You should have received a copy of the GNU General Public License along with this program. If
 * not, see <http://www.gnu.org/licenses/>.
 * </p>
 */
package net.naonedbus.schedule;

/*
 * #%L
 * Naonedbus-server
 * %%
 * Copyright (C) 2010 - 2013 Naonedbus
 * %%
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as
 * published by the Free Software Foundation, either version 3 of the 
 * License, or (at your option) any later version.
 * 
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 * 
 * You should have received a copy of the GNU General Public 
 * License along with this program.  If not, see
 * <http://www.gnu.org/licenses/gpl-3.0.html>.
 * #L%
 */


import net.naonedbus.schedule.job.CommonJob;

/**
 * Classe en charge d'appeler les jobs. Cette "étape" dans l'appel d'un job est nécessaire pour
 * éviter le problème de création de Proxy par l'annotation {@code @Transactionnal}.
 * @author Benoît
 * @version $Revision$ $Date$
 */
public class CommonScheduler
{
    /**
     * Job à exécuter.
     */
    private CommonJob job;

    /**
     * Méthode appelée lors du déclenchement du scheduler.
     */
    public void schedule()
    {
        this.job.execute();
    }

    /**
     * Setter pour job.
     * @param job Le job à écrire.
     */
    public void setJob(final CommonJob job)
    {
        this.job = job;
    }
}
